+++ .atomist/handlers/event/TypeScriptEventHandler.ts
@@ -1,7 +1,8 @@
-import { HandleEvent, DirectedMessage, ChannelAddress, Plan } from '@atomist/rug/operations/Handlers';
-import { Match } from '@atomist/rug/tree/PathExpression';
-import { EventHandler, Tags } from '@atomist/rug/operations/Decorators';
-import { Tag } from "@atomist/cortex/stub/Tag";
+import { EventHandler, Tags } from "@atomist/rug/operations/Decorators";
+import { ChannelAddress, DirectedMessage, HandleEvent, Plan } from "@atomist/rug/operations/Handlers";
+import { Match } from "@atomist/rug/tree/PathExpression";
+
+import { Tag } from "@atomist/cortex/Tag";
 
 /**
  * A sample TypeScript event handler used by AddTypeScriptEventHandler.
@@ -9,9 +10,9 @@ import { Tag } from "@atomist/cortex/stub/Tag";
 @EventHandler("TypeScriptEventHandler", "sample TypeScript event handler used by AddTypeScriptEventHandler", "/Tag()")
 @Tags("documentation")
 export class TypeScriptEventHandler implements HandleEvent<Tag, Tag> {
-    handle(event: Match<Tag, Tag>): Plan {
-        let root: Tag = event.root();
-        let message = new DirectedMessage(`${root.nodeName()} event: ${root.name}`, new ChannelAddress("general"));
+    public handle(event: Match<Tag, Tag>): Plan {
+        const root: Tag = event.root();
+        const message = new DirectedMessage(`${root.nodeName()} event received`, new ChannelAddress("#general"));
         return Plan.ofMessage(message);
     }
 }