+++ .atomist/tests/handlers/command/TypeScriptCommandHandlerSteps.ts
@@ -1,17 +1,17 @@
-import { Given, When, Then, HandlerScenarioWorld, CommandHandlerScenarioWorld } from "@atomist/rug/test/handler/Core"
-import { ResponseMessage } from '@atomist/rug/operations/Handlers'
+import { ResponseMessage } from "@atomist/rug/operations/Handlers";
+import { CommandHandlerScenarioWorld, Given, HandlerScenarioWorld, Then, When } from "@atomist/rug/test/handler/Core";
 
-Given("nothing", f => { });
+Given("nothing", (f) => { return; });
 
 When("the TypeScriptCommandHandler is invoked", (world: HandlerScenarioWorld) => {
-    let w: CommandHandlerScenarioWorld = world as CommandHandlerScenarioWorld;
-    let handler = w.commandHandler("TypeScriptCommandHandler");
+    const w: CommandHandlerScenarioWorld = world as CommandHandlerScenarioWorld;
+    const handler = w.commandHandler("TypeScriptCommandHandler");
     w.invokeHandler(handler, {});
 });
 
 Then("you get the right response", (world: HandlerScenarioWorld) => {
-    let w: CommandHandlerScenarioWorld = world as CommandHandlerScenarioWorld;
+    const w: CommandHandlerScenarioWorld = world as CommandHandlerScenarioWorld;
     const expected = "Successfully ran TypeScriptCommandHandler: default value";
     const message = (w.plan().messages[0] as ResponseMessage).body;
-    return message == expected;
+    return message === expected;
 });